requirements.txt
selenium
webdriver-manager
python-dotenv
requests
aiogram
uvicorn
fastapi